[rigging] Lift ConcurrencyLimitInterceptor into shared rpc module
iris and finelog each carried a copy of ConcurrencyLimitInterceptor.
Move the canonical (async-only, asyncio.Semaphore-backed, cancellation-
safe) version to rigging.rpc and update both consumers to import from
there. Drops the buggy finelog copy that still used blocking
threading.Semaphore acquire in the async path.

Adds connect-python and pytest-asyncio to rigging's deps. Test coverage
for the interceptor moves to lib/rigging/tests/test_rpc.py; iris keeps
the RequestTimingInterceptor tests (iris-specific stats integration)
and finelog keeps the SlowRpcInterceptor tests.

Per-PR comment on #5925.
