+++ lib/rigging/pyproject.toml
@@ -7,6 +7,7 @@ name = "marin-rigging"
 version = "0.99"
 requires-python = ">=3.11,<3.13"
 dependencies = [
+    "connect-python>=0.9.0",
     "fsspec>=2024.0.0",
     "gcsfs>=2024.0.0",
     "s3fs>=2024.0.0",
@@ -15,6 +16,7 @@ dependencies = [
 [dependency-groups]
 dev = [
     "pytest>=8.3.2",
+    "pytest-asyncio",
     "pytest-timeout",
 ]
